requirements.txt
Django==3.1.1
dash==1.13.4
dash-core-components==1.10.1
dash-html-components==1.0.3
dash-bootstrap-components==0.10.3
pandas==1.0.5
numpy==1.19.0
dash-table==4.8.1
plotly==4.8.2
django-plotly-dash==1.4.2
channels==2.4.0
channels-redis==3.1.0
Pillow==7.0.0
django-bootstrap4==2.2.0
xlrd==1.2.0
django-tinymce==3.2.0
dpd-components==0.1.0
dpd-static-support==0.0.5
python-dotenv==0.15.0
